Fixed windows build.
git-svn-id: https://aug.svn.sourceforge.net/svnroot/aug/trunk@457 a3a97f7c-8e16-0410-9893-1d5f7d0f8f6e

+++ src/c/augutil/win32/ptimer.c
@@ -64,7 +64,7 @@ aug_elapsed(aug_ptimer_t ptimer, struct timeval* tv)
 
     now.QuadPart /= ptimer->freq_.QuadPart;
 
-    tv->tv_sec = now.QuadPart / 1000000;
-    tv->tv_usec = now.QuadPart % 1000000;
+    tv->tv_sec = (long)(now.QuadPart / 1000000);
+    tv->tv_usec = (long)(now.QuadPart % 1000000);
     return tv;
 }
